+++ R/SHAP_funcs.R
@@ -24,7 +24,6 @@
 #'
 #' @import data.table
 #' @import xgboost
-#' @importFrom lightgbm lgb.Dataset lgb.train
 #' @importFrom stats cutree dist hclust predict lm na.omit
 #'
 #' @export shap.values
@@ -93,7 +92,6 @@ shap.values <- function(xgb_model,
 #'
 #' @example R/example/example_fit_summary.R
 #' @example R/example/example_categorical.R
-#' @example R/example/example_lightgbm.R
 #'
 shap.prep <- function(xgb_model = NULL,
                       shap_contrib = NULL, # optional to directly supply SHAP values
@@ -319,6 +317,36 @@ shap.plot.summary.wrap2 <- function(shap_score, X, top_n, dilute = FALSE){
   shap.plot.summary(shap_long2, dilute = dilute)
 }
 
+# SHAP importance ----------------------------------------------------------
+
+#' Variable importance as measured by mean absolute SHAP value.
+#'
+#' @param data_long a long format data of SHAP values from
+#'   \code{\link{shap.prep}}
+#' @param names_only If \code{TRUE}, returns variable names only.
+#' @param top_n How many variables to be returned?
+#'
+#' @return returns \code{data.table} with average absolute SHAP
+#' values per variable, sorted in decreasing order of importance.
+#'
+#' @export
+#'
+#' @examples
+#'
+#' shap.importance(shap_long_iris)
+#'
+#' shap.importance(shap_long_iris, names_only = 1)
+#'
+shap.importance <- function(data_long, names_only = FALSE, top_n = Inf) {
+  out <- data_long[, list(mean_abs_shap = mean_value[1]), by = "variable"]
+  out <- out[order(mean_abs_shap, decreasing = TRUE)]
+  n <- min(nrow(out), top_n)
+
+  if (isTRUE(names_only)) {
+    out <- as.character(out[["variable"]])
+  }
+  out[seq_len(n)]
+}
 
 # Dependence plot  --------------------------------------------------------
 
@@ -404,7 +432,9 @@ plot.label <- function(plot1, show_feature){
 #'   that feature. y is default to x, if y is not provided, just plot the SHAP
 #'   values of x on the y-axis
 #' @param color_feature which feature value to use for coloring, color by the
-#'   feature value
+#'   feature value. If "auto", will select the feature "c" minimizing the
+#'   variance of the shap value given x and c, which can be viewed as a
+#'   heuristic for the strongest interaction.
 #' @param data_int the 3-dimention SHAP interaction values array. if `data_int`
 #'   is supplied, y-axis will plot the interaction values of y (vs. x).
 #'   `data_int` is obtained from either \code{predict.xgb.Booster} or
@@ -414,12 +444,16 @@ plot.label <- function(plot1, show_feature){
 #'   20% of the data. As long as dilute != FALSE, will plot at most half the
 #'   data
 #' @param smooth optional to add a _loess_ smooth line, default to TRUE.
-#' @param size0 point size, default to 1 of nobs<1000, 0.4 if nobs>1000
+#' @param size0 point size, default to 1 if nobs<1000, 0.4 if nobs>1000
 #' @param add_hist whether to add histogram using \code{ggMarginal}, default to
 #'   TRUE. But notice the plot after adding histogram is a `ggExtraPlot` object
 #'   instead of `ggplot2` so cannot add `geom` to that anymore. Turn the
 #'   histogram off if you wish to add more `ggplot2` geoms
 #' @param add_stat_cor add correlation and p-value from `ggpubr::stat_cor`
+#' @param alpha point transparancy, default to 1 if nobs<1000 else 0.6
+#' @param jitter_height amount of vertical jitter (see hight in \code{geom_jitter})
+#' @param jitter_width amount of horizontal jitter (see width in \code{geom_jitter}). Use values close to 0, e.g. 0.02
+#' @param ... additional parameters passed to \code{geom_jitter}
 #'
 #' @export shap.plot.dependence
 #'
@@ -438,17 +472,29 @@ shap.plot.dependence <- function(
   smooth = TRUE,
   size0 = NULL,
   add_hist = FALSE,
-  add_stat_cor = FALSE
+  add_stat_cor = FALSE,
+  alpha = NULL,
+  jitter_height = 0,
+  jitter_width = 0,
+  ...
   ){
   if (is.null(y)) y <- x
-  data0 <- data_long[variable == y,.(variable, value)] # the shap value to plot for dependence plot
+  data0 <- data_long[variable == y, .(variable, value)] # the shap value to plot for dependence plot
   data0$x_feature <- data_long[variable == x, rfvalue]
-  if (!is.null(color_feature)) data0$color_value <- data_long[variable == color_feature, rfvalue]
+
+  # Note: strongest_interaction can return NULL if there is no color feature available
+  # Thus, we keep this condition outside the next condition
+  if (!is.null(color_feature) && color_feature == "auto") {
+    color_feature <- strongest_interaction(X0 = data0, Xlong = data_long)
+  }
+  if (!is.null(color_feature)) {
+    data0$color_value <- data_long[variable == color_feature, rfvalue]
+  }
   if (!is.null(data_int)) data0$int_value <- data_int[, x, y]
 
   nrow_X <- nrow(data0)
   if (is.null(dilute)) dilute = FALSE
-  if (dilute!=0){
+  if (dilute != 0){
     dilute <- ceiling(min(nrow(data0)/10, abs(as.numeric(dilute))))
     # not allowed to dilute to fewer than 10 obs/feature
     set.seed(1234)
@@ -456,40 +502,60 @@ shap.plot.dependence <- function(
   }
 
   # for dayint, reformat date
-  if (x == 'dayint'){
-    data0[, x_feature:= as.Date(data0[,x_feature], format = "%Y-%m-%d",
+  if (x == "dayint"){
+    data0[, x_feature:= as.Date(data0[, x_feature], format = "%Y-%m-%d",
                                 origin = "1970-01-01")]
   }
-  if (is.null(size0)) size0 <- if(nrow(data0)<1000L) 1 else 0.4
-  plot1 <- ggplot(data = data0,
-                  aes(x = x_feature,
-                      y = if (is.null(data_int)) value else int_value,
-                      color = if (!is.null(color_feature)) color_value else NULL))+
-    geom_point(size = size0, alpha = if(nrow(data0)<1000L) 1 else 0.6)+
+
+  if (is.null(size0)) {
+    size0 <- if (nrow(data0) < 1000L) 1 else 0.4
+  }
+
+  if (is.null(alpha)) {
+    alpha <- if (nrow(data0) < 1000L) 1 else 0.6
+  }
+  plot1 <- ggplot(
+    data = data0,
+    aes(x = x_feature,
+        y = if (is.null(data_int)) value else int_value,
+        color = if (!is.null(color_feature)) color_value else NULL)
+    ) +
+    geom_jitter(
+      size = size0,
+      width = jitter_width,
+      height = jitter_height,
+      alpha = alpha,
+      ...
+    ) +
     labs(y = if (is.null(data_int)) paste0("SHAP value for ", label.feature(y)) else
-      paste0("SHAP interaction values for\n", label.feature(x), " and ", label.feature(y)) ,
+      paste0("SHAP interaction values for\n", label.feature(x), " and ", label.feature(y)),
          x = label.feature(x),
          color = if (!is.null(color_feature))
            paste0(label.feature(color_feature), "\n","(Feature value)") else NULL) +
     scale_color_gradient(low="#FFCC33", high="#6600CC",
                          guide = guide_colorbar(barwidth = 10, barheight = 0.3)) +
     theme_bw() +
-    theme(legend.position="bottom",
-          legend.title=element_text(size=10),
-          legend.text=element_text(size=8))
-    # a loess smoothing line:
-  if(smooth){
-    plot1 <- plot1 + geom_smooth(method = 'loess', color = 'red', size = 0.4, se = F)
+    theme(legend.position = "bottom",
+          legend.title = element_text(size = 10),
+          legend.text = element_text(size = 8))
+
+  # a loess smoothing line:
+  if (smooth) {
+    plot1 <- plot1 +
+      geom_smooth(method = "loess", color = "red", size = 0.4, se = FALSE)
   }
   plot1 <- plot.label(plot1, show_feature = x)
+
   # add correlation
-  if(add_stat_cor){
+  if (add_stat_cor) {
     plot1 <- plot1 + ggpubr::stat_cor(method = "pearson")
   }
 
   # add histogram
-  if(add_hist){
-    plot1 <- ggExtra::ggMarginal(plot1, type = "histogram", bins = 50, size = 10, color="white")
+  if (add_hist) {
+    plot1 <- ggExtra::ggMarginal(
+      plot1, type = "histogram", bins = 50, size = 10, color = "white"
+    )
   }
 
   plot1
@@ -692,7 +758,6 @@ shap.plot.force_plot_bygroup <- function(
   return(p)
 }
 
-
 # miscellaneous
 if(getRversion() >= "2.15.1")  {
   utils::globalVariables(c(".", "rfvalue", "value","variable","stdfvalue",
@@ -700,5 +765,6 @@ if(getRversion() >= "2.15.1")  {
                            "int_value", "color_value",
                            "new_labels","labels_within_package",
                            "..var_cat",
-                           "group", "rest_variables", "clusterid", "ID", "sorted_id", "BIAS"))
+                           "group", "rest_variables", "clusterid", "ID", "sorted_id", "BIAS",
+                           "cond_mean", "mean_abs_shap"))
 }